+++ index.html
@@ -11,7 +11,10 @@
       rel="stylesheet"
       href="./src/components/transaction-form/transactionform.css"
     />
-    <link rel="stylesheet" href="./src/components/transactionList.css" />
+    <link
+      rel="stylesheet"
+      href="./src/components/transaction-list/transactionlist.css"
+    />
     <title>Document</title>
   </head>
   <body>
