+++ src/components/transaction-list/transactionlist.css
@@ -121,7 +121,7 @@
   display: flex;
   flex-direction: column;
   gap: 16px;
-  margin-bottom: 40px;
+  margin-top: 40px;
 }
 
 .day-header {
