+++ src/observers/TransactionListObserver.js
@@ -1,12 +1,17 @@
-import { renderTransactionList } from "../components/transactionListView.js";
+import { renderTransactionList } from "../components/transaction-list/TransactionListView.js";
+import ItemsState from "../store/ItemsState.js";
 
-export class TransactionListObserver {
-  constructor(itemsState) {
-    this.itemsState = itemsState;
-    itemsState.subscribe(this);
+class TransactionListObserver {
+  constructor(ItemsState) {
+    this.ItemsState = ItemsState;
+    ItemsState.subscribe(this);
   }
 
   update() {
     renderTransactionList();
   }
+}
+
+export function addTransactionListObservers() {
+  new TransactionListObserver(ItemsState);
 }
